feat: Area Calculator Program
A method named area with one double parameter named radius.The method needs to return a double value that represents the area of a circle. Another overloaded method with 2 parameters x and y (both doubles), where x and y represent the sides of a rectangle. The method needs to return an area of a rectangle. If either or both parameters is/are a negative return -1.0 to indicate an invalid value.