Feat: A program that determines how much paint is needed and has to be bought
A method named getBucketCount with 4 parameters. The first parameter should be named width of type double. This parameter represents the width of the wall. The second parameter should be named height of type double. This parameter represents the height of the wall. The third parameter should be named areaPerBucket. This parameter represents the area that can be covered with one bucket of paint. The fourth parameter should be named extraBuckets. This parameter represents the bucket count that Bob has at home.

Another overloaded method named getBucketCount with 3 parameters namely width, height, and areaPerBucket (all of type double).

Another overloaded method named getBucketCount with 2 parameters namely, area and areaPerBucket (both of type double).