Feat: Program to read the numbers entered in the console and printing the minimun and maximun numbers when entering an invalid value.
The program reads the numbers from the console entered by the user and print the minimum and maximum number the user has entered.
Before the user enters the number, the message is printed - Enter number:
If the user enters an invalid number, break out of the loop and print the minimum and maximum number.