Feat: A program that calculates the sum and the average of the entered number
A method called inputThenPrintSumAndAverage that does not have any parameters. The method should not return anything (void) and it needs to keep reading int numbers from the keyboard. When the user enters something that is not an int then it needs to print a message in the format "SUM = XX AVG = YY".